requirements.txt
attr==0.3.2
attrs==21.2.0
faiss==1.5.3
faiss_gpu==1.7.2
matplotlib==3.5.3
numpy==1.21.5
packaging==21.3
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.9.0
SQLAlchemy==1.4.31
sync_batchnorm==0.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
yacs==0.1.8
